# crates/full-node/sov-api-spec/openapi-v3.yaml
...
openapi: 3.0.2
info:
  title: Sovereign Full Node API
  version: 0.1.0
  description: |-
    This is the REST API for the Sovereign SDK full node.
  contact:
    name: Sovereign Labs
    url: https://github.com/Sovereign-Labs/sovereign-sdk-wip
    email: info@sovereign.xyz
  license:
    name: Apache-2.0
    url: https://github.com/Sovereign-Labs/sovereign-sdk-wip/blob/nightly/LICENSE.md
servers:
  - url: http://localhost:12346
    description: Localhost
security: []
paths:
...
  /rollup/schema:
    get:
      summary: Get schema
      description: >
        Returns the current schema used by the rollup.

        This information can be used by web3 SDK and external tools as a description to serialize objects
        as well as determine if the version of the rollup has changed.
      operationId: rollup_schema
      tags:
        - Rollup
      responses:
        "200":
          $ref: "#/components/responses/Schema"
        "400":
          $ref: "#/components/responses/ApiErrorResponse"
...
  /rollup/addresses/{address}/dedup:
    get:
      summary: Deduplicate an address
      description: >
        This endpoint retireves deduplication information needed for building transactions.

        The exact shape of the data in the response depends on the implementation of the rollup.

      operationId: address_dedup
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
          description: The address to retrieve deduplication information about
      tags:
        - Rollup
      responses:
        "200":
          $ref: "#/components/responses/Dedup"
        "400":
          $ref: "#/components/responses/ApiErrorResponse"
...
components:
...
  responses:
...
    Dedup:
      description: Successfully calculated dedup details for the provided address.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AnyJsonValue"
    Schema:
      description: Schema of the rollup.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AnyJsonValue"
...
    ApiErrorResponse:
      description: The request sent was invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    AnyJsonValue:
      oneOf:
        - type: string
          title: String
        - type: number
          title: Number
        - type: boolean
          title: Boolean
        - type: array
          title: Array
          items:
            x-stainless-any: true
        - type: object
          title: Object
          additionalProperties: true
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code related to this error
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - status
        - message
        - details
...